pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "aiortc>=1.14.0",
    "eventlet>=0.40.3",
    "faster-whisper>=1.2.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "python-socketio>=5.14.3",
    "requests>=2.32.5",
    "simple-websocket>=1.1.0",
]
